// src/app/api/model/journalEntryTransactionItem.ts
/**
 * Apache Fineract REST API
 *
 * Contact: dev@fineract.apache.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CurrencyItem } from './currencyItem';
import { EnumOptionType } from './enumOptionType';
import { TransactionDetails } from './transactionDetails';


export interface JournalEntryTransactionItem { 
    amount?: number;
    comments?: string;
    createdByUserId?: number;
    createdByUserName?: string;
    createdDate?: string;
    currency?: CurrencyItem;
    entityId?: number;
    entityType?: EnumOptionType;
    entryType?: EnumOptionType;
    externalAssetOwner?: string;
    glAccountCode?: string;
    glAccountId?: number;
    glAccountName?: string;
    glAccountType?: EnumOptionType;
    id?: number;
    manualEntry?: boolean;
    officeId?: number;
    officeName?: string;
    officeRunningBalance?: number;
    organizationRunningBalance?: number;
    referenceNumber?: string;
    reversed?: boolean;
    runningBalanceComputed?: boolean;
    submittedOnDate?: string;
    transactionDate?: string;
    transactionDetails?: TransactionDetails;
    transactionId?: string;
}